+++ test/runtests.jl
@@ -3,6 +3,7 @@ module TestingLinearInterpolatorsInterpolations
 using LazyAlgebra, TwoDimensional
 
 using LinearInterpolators
+using LinearInterpolators: getcoefs, limits
 
 using Printf
 using Test
@@ -26,6 +27,13 @@ shortname(::Type{T}) where {T} = shortname(string(T))
 shortname(str::AbstractString) =
     shortname(match(r"([_A-Za-z][_A-Za-z0-9]*)([({]|$)", str))
 
+struct CustomCoordinate{T<:Real}
+    val::T
+end
+
+LinearInterpolators.convert_coordinate(T::Type, c::CustomCoordinate) =
+    convert(T, c.val)
+
 kernels = (RectangularSpline(), LinearSpline(), QuadraticSpline(),
            CubicSpline(), CatmullRomSpline(), KeysSpline(-0.4),
            MitchellNetravaliSpline(), LanczosKernel(4), LanczosKernel(6))
@@ -39,6 +47,16 @@ xsub = x[1:sub:end];
 y = cos.(2 .* x .* (x .+ 2));
 ysub = y[1:sub:end];
 t = range(1, stop=length(xsub), length=length(x));
+
+@testset "`getcoefs` method" begin
+    T = Float64
+    ker = CatmullRomSpline(T)
+    len = 4
+    lim = limits(ker, len)
+    for x in (0, .1, -1.2)
+        @test getcoefs(ker, lim, x) == getcoefs(ker, lim, CustomCoordinate(x))
+    end
+end
 
 @testset "TabulatedInterpolators" begin
     tol = 1e-14