+++ NEWS.md
@@ -1,5 +1,20 @@
 # User visible changes for LinearInterpolators
 
+
+## Version 0.1.7
+
+- Extend `getcoefs` to cope with other coordinate types.  The coordinate type
+  is converted on-the-fly if it is not axactly the floating-point type of the
+  kernel.  The method `convert_coordinate` can be specilaized to extend this
+  conversion to non-standard numeric types.  This is automatically done for
+  coordinates of type `Unitful.Quantity` when the `Unitful` package is loaded.
+
+
+## Version 0.1.5
+
+- Bug fixed.
+
+
 ## Version 0.1.3
 
 - `SparseInterpolator(T,ker,...)` and